+++ requirements.txt
@@ -1,6 +1,7 @@
 beautifulsoup4==4.12.3
 chromadb==0.4.24
 mattsollamatools==0.0.21
+langchain
 ollama==0.1.8
 python_magic==0.4.27
 Requests==2.31.0
